Implement read stacking filter
Track each side's interval containing the reads as well as the
breakpoint window, and discard the group if either's width is
(with fuzz 1) no longer than the usual read length. In other words,
discard if all the reads supporting one or other side of the breakpoint
are "stacked" at the same genomic location.