Implement supplementary-only group filter
Discard the group if it is supported only by SUPPLEMENTARY reads.
As the SAM specification requires (for each extant read) exactly one
alignment record with neither SECONDARY nor SUPPLEMENTARY set, in fact
we filter based on primary alignments with neither of those flags set.

As we look in depth at only one read out of the pair, we use the flags
of the observed read and assume that the mate is similarly primary /
supplementary. This is good enough as the intended behaviour of pairing
in the presence of supplementary alignments is unclear in any case --
see samtools/hts-specs#53 et al.