Fix near_mate check to fire only on the same chromosome
The Sanger CGP-added near_mate check discards reads based on their
and their mate's positions even if the read and its mate are on
different chromosomes. This corrects that.

The check probably mostly duplicates the later apparent_insertion
check anyway, but we retain it to stay more comparable to Sanger CGP.