Reimplement swamped filter by counting discordant pairs directly
Accumulate discordant reads in the main read-scanning loop rather
than counting only reads that survive to form other rearr_groups.

Add a parameter for the threshold for applying the swamping filter,
but don't yet connect it to a command line option.

Considering only reads in the main (tumour) sample, discard a group
if there are more than or equal to twice as many discordant read pairs
(thus supporting other structural variants) overlapping with either read
mapping window as there are read pairs supporting the group -- unless
the group is a retrotransposon.

Discordant pairs that are forward/reverse on the same chromosome and only
slightly larger than insert size (defined as <500bp) must not be counted.

Groups that have no supporting reads at all in the main sample will fail
the swamping filter (even if there are also zero nearby discordant reads);
report these separately as a "Not in main sample" filter.

Bump version number to -ww9.